Adjust test files
